% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_metadata.R
\name{print.omicidx_qhits}
\alias{print.omicidx_qhits}
\title{display hits from omicidx query}
\usage{
\method{print}{omicidx_qhits}(x, ...)
}
\arguments{
\item{x}{instance of `omicidx_qhits`}

\item{\dots}{not used}
}
\description{
display hits from omicidx query
}
