% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_metadata.R
\name{query_metadata}
\alias{query_metadata}
\title{query omicidx API for studies based on lucene query}
\usage{
query_metadata(luceneq, cursor_string = "my_cursor")
}
\arguments{
\item{luceneq}{character(1) query in lucene syntax}

\item{\dots}{passed to omicidxClientR::ExperimentsSraExperimentsGet}
}
\description{
query omicidx API for studies based on lucene query
}
\examples{
x = query_metadata("cancer")
str(httr::response(x))
}
